Removed un-necesary pre-processing from the direct ARCADE interface.
The table and description finding function in inbox.py is clever
enough to find the right table and description in the arcade output,
even with all the user interface and odd escape characters at the top
of the input. Also, use the last timetable, rather than the first to
ignore exams.  Possible future addition perhaps?